Fix object-dtype crash when training on a single-performance dataset
np.array([one_2d_array], dtype=object) collapses into a 2D object
array where every cell is a Python float, so slices from it inherit
dtype=object and Keras rejects them with "Invalid dtype: object".
Explicitly passing dtype=np.float32 to the np.array() calls in
PredictiveMusicMDRNN.train() forces the correct type regardless of
how the corpus was stored.

The existing test suite only exercises multi-performance datasets
(heterogeneous array lengths), where numpy stores each performance
as an intact float32 array inside the object array, hiding the bug.

+++ impsy/mdrnn.py
@@ -249,8 +249,8 @@ class PredictiveMusicMDRNN(object):
             callbacks.append(early_stopping_callback)
 
         # Do the data scaling in here.
-        X = np.array(X) * SCALE_FACTOR
-        y = np.array(y) * SCALE_FACTOR
+        X = np.array(X, dtype=np.float32) * SCALE_FACTOR
+        y = np.array(y, dtype=np.float32) * SCALE_FACTOR
 
         ## print out stats.
         print("Number of training examples:")
